\name{lattice}
\alias{lattice}
\title{Lattice of complex numbers}
\description{
Returns a lattice of numbers generated by a  given complex basis.
}
\usage{
lattice(p,n)
}
\arguments{
  \item{p}{Complex vector of length two giving a basis for the lattice}
  \item{n}{size of lattice}
}
\author{Robin K. S. Hankin}
\examples{
 lattice(c(1+10i,100+1000i),n=2)
plot(lattice(c(1+1i,1.1+1.4i),5))
}
\keyword{math}
