\name{nome}
\alias{nome}
\alias{nome.k}
\title{Nome in terms of m or k}
\description{
  Calculates the nome in terms of either \eqn{m}{m} (\code{nome()})
  or \eqn{k}{k} (\code{nome.k()}).
}
\usage{
nome(m)
nome.k(k)
}
\arguments{
  \item{m}{Real parameter}
  \item{k}{Real parameter with \eqn{k=m^2}{k=m^2}}
}
\author{Robin K. S. Hankin}
\examples{
3+3
  }
}
\keyword{math}
